refactor(cli): move merge functions into own module (#171)
